+++ tests/util/mod.rs
@@ -1,9 +1,18 @@
-use std::sync::{Arc, Once, ONCE_INIT};
+use std::sync::{
+    Arc,
+    Once,
+    ONCE_INIT,
+};
 
 use error_chain::ChainedError;
-use jni::{AttachGuard, InitArgsBuilder, JNIEnv, JNIVersion, JavaVM};
 use jni::errors::Result;
-
+use jni::{
+    AttachGuard,
+    InitArgsBuilder,
+    JNIEnv,
+    JNIVersion,
+    JavaVM,
+};
 
 pub fn jvm() -> &'static Arc<JavaVM> {
     static mut JVM: Option<Arc<JavaVM>> = None;
@@ -17,8 +26,8 @@ pub fn jvm() -> &'static Arc<JavaVM> {
             .build()
             .unwrap_or_else(|e| panic!("{}", e.display_chain().to_string()));
 
-        let jvm = JavaVM::new(jvm_args)
-            .unwrap_or_else(|e| panic!("{}", e.display_chain().to_string()));
+        let jvm =
+            JavaVM::new(jvm_args).unwrap_or_else(|e| panic!("{}", e.display_chain().to_string()));
 
         unsafe {
             JVM = Some(Arc::new(jvm));
@@ -29,12 +38,13 @@ pub fn jvm() -> &'static Arc<JavaVM> {
 }
 
 pub fn attach_current_thread() -> AttachGuard<'static> {
-    jvm().attach_current_thread().expect("failed to attach jvm thread")
+    jvm()
+        .attach_current_thread()
+        .expect("failed to attach jvm thread")
 }
 
 pub fn print_exception(env: &JNIEnv) {
-    let exception_occurred = env.exception_check()
-        .unwrap_or_else(|e| panic!("{:?}", e));
+    let exception_occurred = env.exception_check().unwrap_or_else(|e| panic!("{:?}", e));
     if exception_occurred {
         env.exception_describe()
             .unwrap_or_else(|e| panic!("{:?}", e));